// playlist-service-nest/src/interfaces/playlist.ts
/**
 * This file is auto-generated by nestjs-proto-gen-ts
 */

import { Observable } from 'rxjs';
import { Metadata } from '@grpc/grpc-js';

export namespace playlist {
  export interface PlaylistService {
    play(
      data: google.protobuf.Empty,
      metadata?: Metadata,
      ...rest: any[]
    ): Observable<PlaylistResponse>;
    pause(
      data: google.protobuf.Empty,
      metadata?: Metadata,
      ...rest: any[]
    ): Observable<PlaylistResponse>;
    addSong(
      data: AddSongRequest,
      metadata?: Metadata,
      ...rest: any[]
    ): Observable<PlaylistResponse>;
    next(
      data: google.protobuf.Empty,
      metadata?: Metadata,
      ...rest: any[]
    ): Observable<PlaylistResponse>;
    prev(
      data: google.protobuf.Empty,
      metadata?: Metadata,
      ...rest: any[]
    ): Observable<PlaylistResponse>;
  }
  export interface AddSongRequest {
    title?: string;
    duration?: number;
  }
  export interface PlaylistResponse {
    status?: string;
  }
}
export namespace google {
  export namespace protobuf {
    // tslint:disable-next-line:no-empty-interface
    export interface Empty {}
  }
}